pyproject.toml
[project]
name = "testapi"
version = "0.1.0"
description = ""
authors = [
    {name = "TkachevAV",email = "andreytkach233@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "ruff (>=0.12.4,<0.13.0)",
    "alembic (>=1.16.4,<2.0.0)"
]

[tool.poetry]
packages = [{include = "testapi", from = "src"}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.12.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
